chore[Op#40437]: let add comment and edit comment wait for load
activities_tab.{add_comment,edit_comment} waits for comment load so that subsequent assertions don't have to.
Further, this makes things more consistent.

There is a lag on stimulus controller loading, so a sleep time of 1 is adopted "formally" hence TODO removed as the
obvious solution. Should there be a need to increase the sleep time, then we should find an alternative way