attempt to stabilize flickering relations spec
The before block within the context was only executed after the browser already started visiting the page. So only in case the before block
was then executed quickly enough was the queried for relation actually part of the rendered relation tab.
The fix is to execute the relation creation before visiting the relations tab.