Fix time dependent test
Test is dependent on the current time: recurring meetings are created
from 31-12-2024 and the expectations to have 3 occurrences showing up
fail if current time is after 01-01-2025.

Use fixed current time using `travel_to`, but for one test it had to be
in a `before_all` block because some objects were created with
`shared_let`, and modifying them then generated an error because of the
journal creation dates not being in order. It then also had to be
duplicated in the `before(:each)` block of the test but I'm not sure
why.