Move some tests at the `UpdateService` integration test level
Setting attributes like `parent` and then expecting its own dates to be
updated is more an `UpdateService` concern. The fact that it is handled in
a `SetAttributesService` and `SetScheduleService` to update its parent's
dates is an implementation detail.

This should allow refactoring the `SetScheduleService` and
`SetAttributesService` to handle dates assignment only in the
SetAttributesService.