Change replacement_patterns_defined to require an attribute name
This ensures that the method is only used in contexts where we know
for which attribute we expect a replacement pattern to be defined.

This change makes errors less likely where one would expect that
all replacement patterns are defined for subjects (which is the status
quo as implemented, but not as designed).

Also all code that I've seen so far that calls this method, needed to
validate the subject attribute separately afterwards. This change should
make such code shorter.