Fix initial display of pattern input
This was hardcoded to always be hidden. Fixing this
uncovered a minor issue with the code structure. So far
the SubjectConfigurationForm was responsible for mapping the
backend data to frontend state (e.g. radio buttons). However,
the parameter we needed to pass here needed to be defined before
the form could be defined. For now this was solved by making this
decision available as a class method on the form.