Update patterns by changing subject-pattern only
This means if there were other (non-subject) keys in the patterns,
we'd not inadvertently overwrite them.

This includes a change to the Patterns::Collection:
* The subject pattern is now considered optional
    * This is in line with most code ignoring it if it's disabled anyways,
      there was never a strong expectation to always have a subject pattern
* The patterns-collection itself will now always be present on a type, it will
  never be nil (but it can be empty)

The latter change means that we can get rid of some nil-checks, and thus we don't
need to differentiate between "create" cases (when there are no patterns yet) and
"update" cases (there are patterns and we change them). Everything is an update to
the patterns.

This is consistent with other enumerable relations and attributes attached to a
record. E.g. one wouldn't perform a nil-check on a has_many relation.

This also means that invalid assignments, such as

    type.patterns = "Bob"

will now not set patterns to nil, but to the null object (an empty collection)
as well.