Fix WP creation with required hierarchy CFs
In some cases, such as creation through the "New Child" button
in the relations tab, we were not able to render a selector
for the hierarchy custom fields and thus rendering the dialog
failed entirely.

We don't yet have a good component for hierarchical custom fields,
so as a quick fix we enable the SelectList components to also
be usable for hierarchy CFs. We should later introduce proper
hierarchical components for them.