Fix flaky test
The test is opening a dialog and setting text in an autocompleter. That
triggers the dropdown of the autocompleter options. But as the dialog is
still in its opening animation, the dropdown is not displayed at the
right place: it is over the input. When the test then tries to click on
that input, it fails because the dropdown is overlapping the input.
That's the flakiness

This commit adds a helper `wait_for_size_animation_completion(selector)`
to wait for the animation of an element to finish. It's used to wait for
the dialog animation to complete. Once the animation is finished, it is
safe to set the text in the input. It triggers the display of the
dropdown as before, but it is now displayed at the right place and the
test can continue without failing.