Prevent editing non-writable fields
So far this was not ensured when the form wanted to bring
all fields into edit state, e.g. when it's a create form.

However, when the subject is automatically generated, we want to
prevent editing it in all situations. Presumably we had no other
unwritable fields visible inside a creation form so far?