Prevent editing unwritable subject in CreateForm
Doing this by directly using the schema that previous code would
have fetched via API.

This seems to semantically make sense, because regardless of how we
render a component, we should make sure to respect its schema.
On the other hand, it feels weird to directly depend on any class
from the API::V3 namespace in a non-API context.