Translate hint about auto-generated subjects
Previously this hint would have always been shown in english,
regardless of the user's selected language. Now it's possible to
add a translation for this message.

I also slightly rephrased the message, so that it's more in line
with the words we use on the configuration pages. E.g. we didn't
use the word "template" there at all.