Fix progress popover features tests
Debounce value was changed in the app part, but was not updated in the
test.

Also, changed the way the field is focused to avoid the select to be
opened (because it could stay opened sometimes). And for input fields
being focused, it required to also move the carte to the end of the
input before clearing it. It was not needed before because we used to
click in the field to give focus, and clicking also moves the caret to
the end (most of the time).