Split FetchService into three services
Now splitting responsibilities into:
* Fetching
* Refreshing
* Exchanging

Where fetching is composed of the other two services, so that
a caller can mostly care about receiving a usable access token
and not the intricacies of refreshing and exchanging tokens that
were obtained.

This is intended to improve readability and understandability of
the code, while keeping things together that need to be done
in concert.