Allow to configure authentication method through UI
This includes a refactoring to our storage contracts,
which so far required to perform all validations of a storage at
once, even though the storage is assembled step by step.

It's now possible to validate only one step of the storage creation,
while still having a contract that validates the storage as a whole.

By default the Create/Update contract will validate the whole storage,
as before, but they can now be instructed to validate using one
of the partial contracts.