Fix flaky spec
Flaky spec was `./modules/costs/spec/features/cost_entries/add_cost_entry_spec.rb:114`
It was failing for instance in https://github.com/opf/openproject/actions/runs/12934263465/job/36078482478

The flakiness is due to the js on the page not working until the
spent_on date field is fully loaded. When the unit cost values are
input, when the js is not working, then the computed "Costs" value is
not updated, and so the test fails.

The fix is simple: wait for the spent_on date field to be loaded with
`page.has_field?(CostEntry.human_attribute_name(:spent_on))`.

Then I refactored things a bit to remove some duplicate code and use
some accessible selectors instead of css selectors.