Fix test
The date picker updates its calendar more frequently than needed. When
non-working days is toggled, the flatpickr is reinitialized. If the date
is selected before this reinitialization is finished, it selects the
date, then the reinitization happens so it displays the current date,
and then we do some assertions on dates that are no longer displayed.
That's why the test was failing.