Refactor the step-by-step storage setup
The idea is that now there is a central class (a Wizard)
managing the steps necessary to guide the user through creation
of a new storage.

This avoids problems such as the update action of step A being
required to prepare some parts of step B already, so that the
view/component for step B can be rendered correctly.

Since the same components are used inside and outside of the
step-by-step wizard, we are explicitly transferring the state on
whether we are currently in a wizard via parameters. This results
in some degree of boilerplate, but is hopefully more clear than
guessing based on context, whether or not we should show the next
step after submission or go back to regular edit mode.