Utilize wizard in StorageViewComponent
The storage view component now has no own "opinion" on the display
order and grouping of steps anymore, but will merely reflect the wizard
configuration. Changes in the wizard are immediately reflected in the UI
as well. Selection of the right components was also extracted into the
storage registries.

Renamings of existing translations were in most cases avoided, by renaming wizard
steps according to existing translations.

This made it necessary to unify the instantiation of all components rendered
inside the storage view, which was relatively straight-forward for most of them.
The OAuthClientFormComponent was challenging, as there was one use case where it
needed to be initialized with a client that was not yet assigned to the storage.