Fix clicking on start date after due date in date picker
When doing so, it was comparing the start date given as tag attribute to
the clicked date. It was comparing a string '2025-01-30' to a date
object.

Now there are fields for start date and due date attributes, and fields
for the actual value as `Date`. The duration field is also cleared in
this situation.

Updated a test to click on a calendar date instead of using the field
directly so that it is covered by tests.