Fix wrong comments/descriptions
The comments did not describe the actual code behaviour
anymore. Rough history:

* The code comment I added was originally introduced
  in commit 629c73c, with a behavioural change
* Then the corresponding code was moved into a service, where
  the previous comment was reinstated, but the new behaviour
  remained

The change in behaviour was apparently due to https://community.openproject.org/projects/openproject/work_packages/33620,
which hints at the possibility that this setting affects not only
oauth (as in "The OAuth 2.0 Authorization Framework"), but most likely
all SSO-like solutions we have (e.g. LDAP). So maybe oauth is intended
to be short for "omniauth".

This commit only rectifies the descriptions. Considering whether
we should change to a safer default and/or rename the setting is
something for the future.