Restore progress preview controller as it was before
The date picker preview controller and the progress preview controller
were abstracted into a super class. It's too hard to make deep
modifications to the date picker preview controller while keeping the
compatibility with the progress preview controller.

For instance, we break the progress values being formatted on blur
without noticing. This commit adds a test just for this.

This commit undoes some changes made in 5429b8b (commit title is
"Extract logic of the progress modal to be reused for the datepicker
modal").

Once the date picker progress controller works as expected, we'll see
which part can be reused and extracted into an abstraction.