Handle dates swapping when editing dates fields directly
This logic was in the date picker component, meaning it was triggered
when picking another date, but not when editing dates in the input
fields directly.

This commit moves the logic to the preview controller, and removes
almost all responsibilities from the date picker component:

- It sends `date-picker:flatpickr-dates-changed` events to the
  controller when new dates are picked
- It receives `date-picker:flatpickr-set-values events from the
  controller to update its dates, mode, and display of non-working days.
  When the update is done, it does it without reinitializing itself.

Some logic still remains in the date picker component for its
initialization which is based on its HTML attributes.

The controller also changed how it gets the dates values after a turbo
morphing: the values are read from HTML inputs once the morph rendering
is finished.

There are also 2 different debounce for the preview: one of 200ms when
the user is typing in the input fields, and one of 0ms when the user is
clicking on a date in the calendar. In that case the debounce is not
needed as the date is set immediately.