Update test to new behavior
If the dates have not changed, the date picker is not reinitialized and
its dates are not set again. So switching "Working-days only" checkbox
no longer shows the current date again and it's not needed for the test
to wait for the current date to be displayed and switch back to the
correct date.