Introduce superclasses for info and form components of storages UI
All of these shared some common patterns that were always repeated,
so it made sense to unify them in a common base class.

For the info components this also affected inclusion of the
StorageViewInformation module, which is why I promoted this module
to be the new superclass for these components.

Closer inspection will probably reveal more commonalities that can
be moved upwards in the class hierarchy.