Allow to configure SSO with fallback to two-way OAuth 2.0
Our code to select how we authenticate mostly accounted for this
already, so it was just a matter of allowing admins to configure it.

We now focus on what's been explicitly configured, so that it is not
necessary to clear the audience or the oauth client/application
just when changing the settings. Otherwise stale configuration
could lead to the wrong authentication to happen contrary to
the configuration.