[#61332] 15.3 Recurring meetings and meetings changes  https://commun… (#17956)
* [#61332] 15.3 Recurring meetings and meetings changes  https://community.openproject.org/work_packages/61332

[#61332] 15.3 Recurring meetings and meetings changes

https://community.openproject.org/work_packages/61332

* add docs-on one-time meetings

* [#61332] 15.3 Recurring meetings and meetings changes  https://community.openproject.org/work_packages/61332

[#61332] 15.3 Recurring meetings and meetings changes

https://community.openproject.org/work_packages/61332

* Update README.md

* adjust links

* Update README.md

Minor changes to phrasing

* Minor

* Update README.md

Changed "the user" or "a user" to the direct second person "you" to make it consistent with the language preceding the list.

* Update README.md

Minor update to specify that dynamic meetings aren't being replaced by one-time or recurring meetings, but that those are different types of dynamic meetings.

* Changes for accuracy and more details

Also updated to take into account the changes to the phrasing of meeting series frequency that will be introduced by https://community.openproject.org/wp/61476

* minor

* Update docs/user-guide/meetings/recurring-meetings/README.md

* fix a heading name

---------

Co-authored-by: as-op <a.sandorf@openproject.com>
Co-authored-by: Parimal Satyal <88370597+psatyal@users.noreply.github.com>