Explicitly store expiration time of access tokens
Previously we didn't consider the expires_in from the
token endpoint response, thus being unable to know whether
a token was expired or not before using it. While we worked
around that by inspecting tokens that seemed readable as JWTs,
the better approach is to memoize the information returned by the
server, as that also allows to know the expiration time of
opaque access tokens.

We'll keep the JWT extraction around as a fallback mechanism,
but prefer reading the info from the token endpoint, if it's present.