Adapt test to manual scheduling being used when appropriate
A work package can be created with `schedule_manually: false` only if it
has children or predecessors. With the API it's not possible to create a
work package with children directly. But it's possible to create a work
package with a parent with a predecessor. In this case, it's possible to
set scheduling mode to automatic.

This commit adds a new spec to cover this case.