Replace the term nextcloud audience with storage audience
The components we are starting to build should be generally usable
for different storages. Before we sink more cost into overly
specific naming, we should generalize to a name that can be used
anywhere, if we want to.

Left out components that would require renaming files for now, to reduce
the conflict potential a bit.