Do not trigger a refresh event on input field change event.
If an input field is changed, still has focus and the user clicks the
submit button, the form will be submitted, but the cost refresh will
also be triggered after the submission. This leads to a UI bug where the
incorrect cost is being shown after the form update. In order to avoid
this glitch, we do not refresh the cost on the input change events, but
only on the keyup event.