Use default scopes as default value, not as placeholder
The scopes defined as placeholders were not set by OpenProject
directly, but by one of the OIDC-related gems. By defining them
explicitly here, we can be sure that those are the scopes really
requested.

UX-wise this also makes it easier to change the scopes from the
default value. For example adding a scope now consists of adding the scope
instead of typing out all the default scopes before adding your scope.