Make some files ready for frozen string literals
During a previous test, added the frozen string literal comment
to all Ruby files, to see how our codebase would react to that.

The test run uncovered a few files that were trying to modify strings
that originate from string literals. This commit is changed some
of them to be safe under frozen string literals and freezes literals
in the affected files.