[#61465] ensure no hover cards on top-right-avatar and within other hover cards
The first feels weird. You should know who you are. The latter is even a
bug, since a hover card will close when you try to open a new one ->
prohibit this from happening.

While on it, I decided to rework the way that a hover card is configured
within the principal renderer. It now gets its own configuration object
instead of piggybacking on the avatar options. It always felt wrong to
cram the hover card options in there, too.