[61535] Can switch to automatic scheduling with indirect predecessors
https://community.openproject.org/wp/61535

A work package can have ancestors having predecessors. These
predecessors are indirect predecessors of the work package and are
involved in the scheduling if all the ancestors up to the predecessor
are automatically scheduled.

When such an indirect predecessor is present, the work package can be
switched to manual scheduling, even if it has no direct predecessors or
children.

It uses the scope `Relation#used_for_scheduling_of(work_package)` to
find all predecessors involved in scheduling. This encompass both direct
and indirect predecessors.