[61894] Fix disappearing fields in date picker when in automatic mode
https://community.openproject.org/wp/61894

In the backend, the date picker component assumes that the work package
is a milestone if the frontend did not send any `work_package[due_date]`
value for "Finish date" field.

Problem is, when in automatic scheduling mode, the "Finish date" field
is disabled, so it's not sent to the backend. The backend then
misleadingly thinks that it's a milestone being edited, and displays
only the "Date" field, as a milestone work package would do. The "Finish
date" and "Duration" fields disappear.

As the "Duration" field is always enabled, the fix is rely on it
(`work_package[duration]`) instead of the "Finish date" field.