Make recurring meeting specs work reliably
The way those specs were written they would not bbe able to
succeed before 10 AM in the morning, because meeting were scheduled
for 10:00, but the specs did not expect a meeting to be scheduled
today.

This fix is a workaround that makes sure we always run the specs
"after 10". A better fix would be to ALSO have specs running
before 10, and expecting the behaviour that's currently
seen as errorneous.