Ensure that we only query origin_user_id when it's missing
Previously we queried the remote storage whenever we exchanged
a token for a user, just to confirm that we still know their
origin_user_id.

Now we will only query the origin_user_id when the remote identity
is new. If we already have a remote identity, we will not try to
update it and will not make additional web requests.