Failed HTTP requests do not raise errors from TOKEN_OBTAINED_EVENT
We will now propagate HTTP errors upwards as failed service results,
top-most they are only logged, but will not raise an error anymore.

We can do this, because a missing remote identity will be shown to
the user as an error anyways. On the other hand raising an error
from the OpenProject::Notifications breaks the code performing
the original token fetching and prevents display of the nice error
message in the UI.