Make sure that remote identity is updated from ConnectionManager
The main intention of switching to the approach of requiring the record
remote identity to be new/unpersisted, was to avoid unnecessary calls to
extract_origin_user_id, which results in a web call. However, the ConnectionManager
already ensures to only call the RemoteIdentities::CreateService is a small number
of cases, so it should be fine to always update when calling it.

We also identified one corner case, where it might be preferable to do that
update from the ConnectionManager: Users could log-in to a storage with
a different account after removing a previous OAuth connection. Without the
update, we'd not change the origin_user_id properly. The risk of that happening
is much lower in an SSO scenario.