Prepare nextcloud storage docs for new authentication method
This is a technical pre-write of the docs. I've decided to
split the instructions per authentication method, because the
SSO-specific instructions need to go into a lot of detail in
configuring third-party systems. They will also be irrelevant for
a large amount of users, because SSO will be a corporate feature.

Proof reading of this draft will certainly be necessary.