Ensure that sorting of locales always works
This intends to fix rarely occuring errors on that line
according to AppSignal:

    comparison of String with :en failed
    Backtrace: app/seeders/basic_data/setting_seeder.rb:72:in 'Array#sort'

I don't know how the symbol ended up in the array, there is possibly
a root cause elsewhere to be found. However, considering that this line
of code already acknowledged that locales are sometimes represented as
a symbol, it is probably sensible to convert the whole array to its
string representation anyways.