Allow access to new file links by creator
For example on the work package create UI, it
is already possible to link files, even though the
WP does not yet exist. In this case it's a usability
issue if the UI elements do not yet work as they should.

As long as there is no associated container, it should be no
issue to assume that the creator has all the necessary permissions
to view a file link. Downloading the associated file is still
limited by permissions on the storage, so this should not
be abusable to retain long-term access to a file.