[62125] Avoid jumping to a different month when setting flatpickr dates
https://community.openproject.org/wp/62125

When setting flatpickr dates with `.setDates`, it jumps the calendar to
the last date so that it's visible on the left side. This is annoying
when clicking on a date on the right side because the calendar view then
changes while it could stay the same.

This commit fixes the issue in two steps:
- first it jumps to the date that has been changed, so that when setting
the start date, the calendar shows the start date instead of showing the
finish date
- then if the month previously displayed was on the right side, it jumps
back to the previously displayed month so that the calendar view stays
the same.

Also it does not jump to different dates if no dates have changed (for
instance when focusing some other fields).