[62012] Avoid resetting calendar when enabling a date field
https://community.openproject.org/wp/62012

When clicking on a date field in single date mode, it used to
reinitialize the flatpickr instance. This happens because it was using a
turbo-stream to update the dialog. When doing so, the node where the
flatpickr is attached is replaced.

We have some logic to avoid this replacement in the preview controller:
it listens to turbo:before-frame-render events and avoids morphing for
tags starting with "opce".

As the logic is already there, this commit uses a turbo-frame instead of
a turbo-stream. This way the flatpickr instance is not replaced and the
calendar view is kept.

The controller had to be updated to alter the work package attributes
with the given params.