Hide SSO+fallback authentication method
While it's useful and works as intended, we can't
configure the same on the openproject_integration in
Nextcloud so far. This means, right now this option
would only make sense for one-way integrations.

The plan is to support this in a future version of
the integration, which is when we can reactivate the
option in OpenProject as well.