Do not switch scheduling mode if it's during a bulk copy
This fixes the failing test which tested a bluk copy of two manually
scheduled work packages being parent and child. The test was failing
because the parent switched to automatic scheduling mode.
spec is `./spec/controllers/work_packages/moves_controller_spec.rb:504`

It uses the `state` present in services inheriting from `BaseCallable`
to pass a `bulk_copy_in_progress` flag which is then checked in the
`UpdateAncestorsService`.