[61979] Switch successor to automatic mode even without rescheduling
https://community.openproject.org/wp/61979

Switching to automatic mode was done in `ScheduleDependency` before. As
the scheduling mode was changed outside of the `SetScheduleService`, it
was not seen as a change and no dependent result was saved and
eventually the scheduling mode change was not persisted.

By moving the scheduling mode change to the `SetScheduleService`, the
dependent results are now correctly saved and the scheduling mode change
is persisted.