[62245] Fix date picker width to 600px
https://community.openproject.org/wp/62245

Text in the banner can make the date picker too wide. The width is now
fixed to 600px to avoid it and let the banner text wrap.

Before the limit was already 600px, but it was only applied to the body.
Now it's applied to the whole content (header, body and footer).