Ensure SSO users do not fallback to ensure_connection
ensure_connection is intended for OAuth2 authentication at storage,
but does not work, if we don't have an OAuth client, like for
SSO authentication.

In the SSO case we can simply try to redirect the user to the correct
URL, because they will most likely be able to login there and
if in doubt, the storage can still refuse their request to visit
the desired path.
Logging in to the storage, can in some cases lead to the user account
being created on the storage, which should fix the main case of failed
authentication in the first place. So hopefully this fallback helps
ensuring users are eventually provisioned on the storage.