[62261] Fix error displayed when switching parent to automatic
https://community.openproject.org/wp/62261

When switching a parent to automatic scheduling mode, the date picker
displayed an error that the dates are read-only when the selected dates
in manual mode are the ones that would be computed by automatic mode.

This was due to system changes being the same as user changes, and as
they were done after, it was not detected.

Fix is to track the attributes directly (maybe using private activemodel
api) and check if the instance changed, in which case it can be
considered a system change.