Seed a separate OAuth application for system API tokens
We want to offer API tokens comparable to the personal access tokens
that a user can have, but on the system-level.

To facilitate code reuse for validating them, the idea is that those
tokens will be associated to a builtin OAuth application. This means
validation of tokens can happen through existing means, because the
tokens will technically just be client credential tokens of that app.