[62362] Show "working days only" checkbox only if date form is shown too
https://community.openproject.org/wp/62362

When clicking the checkbox, it takes all form data and sends it to the
server for rerendering the form with the value updated.

Problem is: when the date form is not displayed and the "no
predecessors" message is displayed instead, there is no form data, so it
sends incomplete data to the server, and the server renders a minimalist
form.

The fix is to hide the checkbox too if the other form elements are hidden.
No checkbox, no click, no problem.