Update test with single date mode
This date was previously skipped because of the way the `:driver` metadata was used.

It was failing because it was not updated to the new single mode, and
the due date was not active, so clicking the date in the calendar was
changing the start date, whereas the test expected the due date to be
set.

The test has been fixed by enabling the due date first. Some additional
checks have been added as well.