Fix SSO storage creation via API
There were several bad assumptions in the previous
implementation. First, the represented is not a storage,
but effectively a Hashie::Mash, so just a fancy hash.
Thus methods such a provider_type_nextcloud? don't exist,
but sadly no error is raised, because `?` methods will just
return false if a key with the same name is missing.

Also on a more general level, it doesn't make sense skip parsing
certain fields depending on the type, because we can't make sure
that the type is being parsed first. Thus we might skip parsing
fields and afterwards realize that we should have parsed them
in the first place.

If we wanted to prevent assigning attributes that we don't want to
assign, it would be the responsibility of SetAttributesService
and contracts to ensure that.