[62487] Fix rescheduling when creating a child through API
https://community.openproject.org/wp/62487

There are two important steps when creating a child through the API:
- Update the ancestors attributes based on the child attributes
- Update the scheduling of work packages that are related to the child

The order is important as the parent will switch to automatic mode if it
has no predecessors and gains its first child. The switch to automatic
mode must occur before rescheduling.

Previous code was buggy because the rescheduling happened first: the
parent was still in manual mode and so its dates were not updated.

Also fixed some other unit tests which were invalid (doing assertions on
the wrong objects.)