Fail when user can't be found
Previously we would have "fallen out" of the
JWT strategy, even though the token was clearly intended
to be handled by this strategy. For my configuration this
led to the final error message being generated by the Basic auth
strategy, leading to a wrong error message.

Backported into 15.4 to avoid conflicts.