Add service accounts
Service accounts are intended to be special users
that are not explicitly manageable through the users UI
(just like "System" and the like), but are intended for
identification of "services", not real people.

The initial scope of this will be to allow authentication
of SCIM clients, that will be able to provision users into
and out of OpenProject. However, the general idea is to
allow using service accounts for generic application-to-application
communication as well. Given proper permissions and scopes (i.e. api_v3),
it would also be possible to use a service account for requests
against our normal APIv3.