Ensure edit state uses restriction
The downside is that if the user has both an edit editor open and one adding a comment, the restriction prevails
so they would not be able to mention non restricted users. Unless they close out the restricted editor. This is mainly
because we are using a singular function to retrieve the mentions- the could be a better way that I didn't yet think about
but my reckoning is this is a rare edge case- and I much prefer the secure by default approach