Fir wrong test conditions
The setup was incorrect: the successor's dates were identical to the predecessor's dates which is not possible, and the milestone had start date different from due date which is not possible either.

Finally the second test was wrong too because when a successor milestone is
automatically scheduled, the date can't be changed and the date picker
is fully disabled.