[62607] Always display relations in the relations tab in a fixed order
https://community.openproject.org/wp/62607

The order of the relation groups was not consistent from one work
package to another. It's because the relations are returned in any order
from the database.

Instead of getting relations from database and then grouping them by
type as they come up in the array, get the types in the order we want
and build up the relation groups from there.

There are no differences between rendering a `RelationGroup` of type
child and a `RelationGroup` of type relation anymore. For this to
happen, the "children" type had to be renamed "child", which is in line
with the Relation::TYPE_CHILD constant.