[62595] Fix date buttons not opening input in Safari in date picker
https://community.openproject.org/wp/62595

Idiomorph is used to morph the date picker dialog content, except the
active element. When clicking the date button in Safari, it does not
become the active element because it's actually a link, and for some
reason links cannot be active in Safari. What becomes active instead is
the closest focusable parent element. It is the whole
`panel-wp-datepicker-dialog--content-tab--dates` as it has `tabindex:
0`. As idiomorph does not change the active element, that's why the date
fields stayed the same.

To fix this, we make pass `ignoreActiveValue = true` to idiomorph only
if the active element is an input. This way the whole date picker form
is always updated when no input is focused.

It also fixes similar issues that existed on "Today" links and "Working
days only" checkbox too.

The `data: { skip_morphing: true }` is some dead code leftover that is
no longer used.