[62641] Keep dates and update duration when toggling "Working days only"
https://community.openproject.org/wp/62641

Previously we tried to keep duration over the dates when updating
"Working days only", also known as `ignore_non_working_days`. The
behavior was a bit surprising and was inconsistent with the fact that
setting a new start date would keep the the finish date and update the
duration.

Now this is consistent: when changing something, we always try keeping
the dates and updating the duration accordingly.