Use job_id from URL to display errors from given job
When giving a url of the first attempt, the script was displaying the
results of the last attempt. That's probably not the intended behavior.

Now it parses the job_id from the url, and if present it will display
the errors from that job.

Also changed the output to display the instructions to checkout the
correct commit at the end of the script, to make it easier to see.