[62891][63152] Recalculate start date when toggling "Working days only"
https://community.openproject.org/wp/62891

For an automatically scheduled successor, when "Working days only" is
turned off and the day before the start date is a non-working day, the
start will shift to an earlier date. Multiple things need to work
correctly:

- The start date must be recalculated and displayed in the date picker.
- The mini calendar of the date picker need to have enabled dates
  updated, as the start date has changed.
- If the due date was not set manually, duration must be kept and due
  date will be recalculated.
- If the due date was set manually, it must not be changed and duration
  will be recalculated.
- If the due date was unset manually, it must remain unset and duration
  must be unset too.

Also the lag calculation was partly incorrect: it was adding non-working
days when "Working days only" was on, and days when "Working days only"
was off. This has been fixed: lag is always about working days
regardless of the successor "Working days only" setting.

https://community.openproject.org/wp/63152

When the browser timezone offset is negative, the enabled dates in the
date picker for an automatically scheduled successor were not correct:
the date before the start date was enabled.

This is now fixed.