[62891] Automatically readjust automatic start date if wrong
There was a bug in previous versions where the lag was all-days when
successor's ignore_non_working_days was true, leading to a start being
too early. The correct behavior for lag is to be working days only, even
if the successor ignores non-working days.

The set attributes service must be able to fix these wrong dates so that
the date picker does not open with an error message saying that the
start date is too early (especially as the start date cannot be changed).